# openapi/data-referee-v1.json
{
  "components": {
    "schemas": {
      "Body_source_file_upload_v1_datasets__dataset_id__source_file_put": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_source_file_upload_v1_datasets__dataset_id__source_file_put",
        "type": "object"
      },
      "DatasetCreate": {
        "properties": {
          "column_mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Column Mapping",
            "type": "object"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "retention_hours": {
            "default": 24,
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Retention Hours",
            "type": "integer"
          },
          "source_type": {
            "default": "csv",
            "maxLength": 50,
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "DatasetCreate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "QualityRunCreate": {
        "properties": {
          "profile_key": {
            "title": "Profile Key",
            "type": "string"
          },
          "profile_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Version"
          }
        },
        "required": [
          "profile_key"
        ],
        "title": "QualityRunCreate",
        "type": "object"
      },
      "RecordBatchCreate": {
        "properties": {
          "batch_id": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Batch Id",
            "type": "string"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/RecordEnvelope"
            },
            "maxItems": 1000,
            "title": "Records",
            "type": "array"
          }
        },
        "required": [
          "batch_id",
          "records"
        ],
        "title": "RecordBatchCreate",
        "type": "object"
      },
      "RecordEnvelope": {
        "properties": {
          "external_record_id": {
            "maxLength": 500,
            "minLength": 1,
            "title": "External Record Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "external_record_id",
          "payload"
        ],
        "title": "RecordEnvelope",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Explainable dataset profiling, validation, classification, and quality exports.",
    "title": "Data Referee",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/datasets": {
      "post": {
        "operationId": "datasets_create_v1_datasets_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Data-Referee-Key"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-client",
            "required": false,
            "schema": {
              "default": "public-ui",
              "title": "X-Data-Referee-Client",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Datasets Create",
        "tags": [
          "Data Referee"
        ]
      }
    },
    "/v1/datasets/{dataset_id}/profile": {
      "get": {
        "operationId": "dataset_profile_v1_datasets__dataset_id__profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Data-Referee-Key"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-client",
            "required": false,
            "schema": {
              "default": "public-ui",
              "title": "X-Data-Referee-Client",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dataset Profile V1 Datasets  Dataset Id  Profile Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dataset Profile",
        "tags": [
          "Data Referee"
        ]
      }
    },
    "/v1/datasets/{dataset_id}/quality-runs": {
      "post": {
        "operationId": "quality_runs_create_v1_datasets__dataset_id__quality_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Data-Referee-Key"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-client",
            "required": false,
            "schema": {
              "default": "public-ui",
              "title": "X-Data-Referee-Client",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityRunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quality Runs Create",
        "tags": [
          "Data Referee"
        ]
      }
    },
    "/v1/datasets/{dataset_id}/record-batches": {
      "post": {
        "operationId": "record_batches_create_v1_datasets__dataset_id__record_batches_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Data-Referee-Key"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-client",
            "required": false,
            "schema": {
              "default": "public-ui",
              "title": "X-Data-Referee-Client",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordBatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Batches Create",
        "tags": [
          "Data Referee"
        ]
      }
    },
    "/v1/datasets/{dataset_id}/source-file": {
      "put": {
        "operationId": "source_file_upload_v1_datasets__dataset_id__source_file_put",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Data-Referee-Key"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-client",
            "required": false,
            "schema": {
              "default": "public-ui",
              "title": "X-Data-Referee-Client",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_source_file_upload_v1_datasets__dataset_id__source_file_put"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source File Upload",
        "tags": [
          "Data Referee"
        ]
      }
    },
...
    "/v1/profiles/{profile_key}/{profile_version}": {
      "get": {
        "operationId": "profile_v1_profiles__profile_key___profile_version__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_key",
            "required": true,
            "schema": {
              "title": "Profile Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "profile_version",
            "required": true,
            "schema": {
              "title": "Profile Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Profile V1 Profiles  Profile Key   Profile Version  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Profile",
        "tags": [
          "Data Referee"
        ]
      }
    },
    "/v1/quality-runs/{run_id}": {
      "get": {
        "operationId": "quality_run_status_v1_quality_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Data-Referee-Key"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-client",
            "required": false,
            "schema": {
              "default": "public-ui",
              "title": "X-Data-Referee-Client",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Quality Run Status V1 Quality Runs  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quality Run Status",
        "tags": [
          "Data Referee"
        ]
      }
    },
    "/v1/quality-runs/{run_id}/dimensions": {
      "get": {
        "operationId": "quality_run_dimensions_v1_quality_runs__run_id__dimensions_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Data-Referee-Key"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-client",
            "required": false,
            "schema": {
              "default": "public-ui",
              "title": "X-Data-Referee-Client",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Quality Run Dimensions V1 Quality Runs  Run Id  Dimensions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quality Run Dimensions",
        "tags": [
          "Data Referee"
        ]
      }
    },
    "/v1/quality-runs/{run_id}/exports/report.html": {
      "get": {
        "operationId": "quality_run_html_report_v1_quality_runs__run_id__exports_report_html_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Data-Referee-Key"
            }
          },
          {
            "in": "header",
            "name": "x-data-referee-client",
            "required": false,
            "schema": {
              "default": "public-ui",
              "title": "X-Data-Referee-Client",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quality Run Html Report",
        "tags": [
          "Data Referee"
        ]
      }
    },
...
  }
}